refactor lists into componenets
